logs page added
